Split routes.py into multiple files
